chatbot/src/backend/python/requirements.txt
faiss-cpu>=1.11.0
sentence-transformers>=4.1.0
langchain-groq>=0.3.2
numpy>=2.2.6
langchain>=0.3.25
uvicorn>=0.34.2
pydantic>=2.11.5
fastapi>=0.115.12
numpy==1.24.4
langchain==0.1.17
langchain-core==0.1.46
langchain-community==0.0.34
langchain-huggingface==0.0.6
faiss-cpu==1.7.4
python-dotenv==1.0.1
